Add files via upload
To find out the MAC address of a network device, such as a Nanode board, you can use the following steps:

1. Connect the Nanode board to the network using an Ethernet cable.
2. Open the Arduino IDE on your computer and create a new sketch.
3. Go to File > Examples > Ethernet > DhcpAddressPrinter.
4. Upload the sketch to the Nanode board.
5. Open the Serial Monitor by going to Tools > Serial Monitor.
6. Make sure the baud rate is set to 9600 and select "Both NL & CR" for the line ending.
7. Wait for the Nanode board to obtain an IP address from the network.
8. The Serial Monitor will display the IP address, subnet mask, and gateway address of the Nanode board.
9. Write down the MAC address displayed in the Serial Monitor. It will be labeled "MAC: xx:xx:xx:xx:xx:xx".

Alternatively, you can use a network scanner tool, such as Advanced IP Scanner or Fing, to scan your network and find the MAC address of the Nanode board. Once you have the IP address of the board, you can scan the network for devices and retrieve their MAC addresses.